refactor: 여러군데에서 사용되는 API를 위해 패키지 구조 변경
- 그룹 목록 호출 API가 메인화면 메뉴와 글쓰기 3단계에서 쓰임
