+++ frontend/rush/src/api/FindCommentsApi.js
@@ -1,5 +1,5 @@
 import axios from "axios";
-import {BACKEND_ADDRESS} from "../../constants/ADDRESS";
+import {BACKEND_ADDRESS} from "../constants/ADDRESS";
 
 const findCommentsApi = (articleId) => {
   return axios.get(BACKEND_ADDRESS + "/comments?article_id=" + articleId)
