+++ frontend/rush/src/components/articleDetail/ArticleDetailPage.js
@@ -2,11 +2,11 @@ import React, {useEffect, useState} from 'react';
 import WindowSize from "../WindowSize";
 import {CommentsBox, DisplayBox, Outside, PostBox} from './Box';
 import ArticleMeta from "./ArticleMeta";
-import findWritingApi from "./FindWritingApi";
+import findWritingApi from "../../api/FindWritingApi";
 import ArticleBody from "./ArticleBody";
 import Comment from "./Comment";
 import CommentWriting from "./CommentWriting";
-import findCommentsApi from "./FindCommentsApi";
+import findCommentsApi from "../../api/FindCommentsApi";
 import {ACCESS_TOKEN} from "../../constants/SessionStorage";
 import {withRouter} from "react-router-dom";
 import {GROUPED, PRIVATE, PUBLIC} from "../../constants/MapType";
