+++ frontend/rush/src/components/home/button/groupMap/GroupList.js
@@ -2,7 +2,7 @@ import React, {useEffect, useState} from 'react';
 
 import {Motion, spring} from 'react-motion';
 import GroupContent from "./GroupContent";
-import findMyGroupsApi from "../../FindMyGroupsApi";
+import findMyGroupsApi from "../../../../api/FindMyGroupsApi";
 import styled from "styled-components";
 
 const GroupSetting = styled.li`
